# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: GitSense API
  description: GitHub 开源生态发现平台 API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发

paths:
  /search:
    post:
      summary: 搜索并推荐相似仓库
      operationId: search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              url_input:
                value:
                  repo_url: "https://github.com/agentgateway/agentgateway"
              owner_repo_input:
                value:
                  repo_url: "agentgateway/agentgateway"
      responses:
        '200':
          description: 搜索成功，返回推荐结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '202':
          description: 仓库已入库，Embedding 正在异步生成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /repos/{owner}/{name}:
    get:
      summary: 获取仓库详情
      operationId: getRepo
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 仓库详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '404':
          description: 仓库不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /repos/{owner}/{name}/recommendations:
    get:
      summary: 获取相似仓库推荐
      operationId: getRecommendations
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: 推荐结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  similar_repositories:
                    type: array
                    items:
                      $ref: '#/components/schemas/SimilarRepository'
        '500':
          description: 内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /repos/{owner}/{name}/ecosystem:
    get:
      summary: 获取技术生态
      operationId: getEcosystem
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 技术生态信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ecosystem'
        '404':
          description: 生态信息不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: 健康检查
      operationId: health
      responses:
        '200':
          description: 服务健康状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok, degraded]
                  database:
                    type: string
                    enum: [connected, disconnected]
                  redis:
                    type: string
                    enum: [connected, disconnected]

/admin/seed:
    post:
      summary: 批量导入 GitHub 仓库
      description: 批量导入仓库并触发异步 Embedding 生成
      operationId: seed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeedRequest'
      responses:
        '200':
          description: 导入结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeedResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    SearchRequest:
      type: object
      required:
        - repo_url
      properties:
        repo_url:
          type: string
          description: "GitHub 仓库 URL 或 owner/name 格式"
          example: "https://github.com/agentgateway/agentgateway"
        limit:
          type: integer
          description: "返回推荐数量，默认 10"
          default: 10
          minimum: 1
          maximum: 50

    Repository:
      type: object
      properties:
        id:
          type: integer
          format: int64
        full_name:
          type: string
          example: "agentgateway/agentgateway"
        owner:
          type: string
        name:
          type: string
        description:
          type: string
        language:
          type: string
        stars:
          type: integer
        topics:
          type: array
          items:
            type: string
        embedding_status:
          type: string
          enum: [pending, generating, done, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    RecommendationReason:
      type: object
      properties:
        embedding_similarity:
          type: number
          format: float64
        topic_similarity:
          type: number
          format: float64
        language_match:
          type: boolean
        common_topics:
          type: array
          items:
            type: string

    SimilarRepository:
      type: object
      properties:
        repository:
          $ref: '#/components/schemas/Repository'
        score:
          type: number
          format: float64
        reason:
          $ref: '#/components/schemas/RecommendationReason'

    Ecosystem:
      type: object
      properties:
        name:
          type: string
          example: "AI Agent Infrastructure"
        repositories:
          type: array
          items:
            $ref: '#/components/schemas/Repository'

    SearchResponse:
      type: object
      properties:
        repository:
          $ref: '#/components/schemas/Repository'
        similar_repositories:
          type: array
          items:
            $ref: '#/components/schemas/SimilarRepository'
        ecosystem:
          $ref: '#/components/schemas/Ecosystem'

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

    SeedRequest:
      type: object
      required:
        - repos
      properties:
        repos:
          type: array
          items:
            type: string
          description: "GitHub 仓库列表（owner/name 格式）"
          example: ["agentgateway/agentgateway", "BerriAI/litellm"]
          maxItems: 100

    SeedResponse:
      type: object
      properties:
        imported:
          type: integer
          description: 成功导入数量
        failed:
          type: array
          items:
            type: string
          description: 失败列表